+++ deticafe/src/main/java/tqs_project/deticafe/model/Type.java
